Change type of count_nodes from int to size_t

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int count_nodes = 0;
+	size_t count_nodes = 0;
 	const list_t *curr = h;
 
 	if (h == NULL)
